// project/web/src/graphql/api/operations.ts
export type CutQuery = {
  __typename?: 'Query';
  cut?: {
    __typename?: 'Cut';
    id: number;
    src: string;
    votesCount: number;
    isVoted: boolean;
    film: { __typename?: 'Film'; id: number; title: string };
  } | null;
  cutReviews: Array<{
    __typename?: 'CutReview';
    id: number;
    contents: string;
    createdAt: string;
    updatedAt: string;
    isMine: boolean;
    user: { __typename?: 'User'; email: string; username: string; profileImage?: string | null };
  }>;
};